Update serialization strategy
All Mobjects are serialized, even if they are never added to the Scene.
Transformations do not propagate to submobjects. Diffs are computed before and
after each Animation, even for Mobjects that aren't in the Scene and even for
Mobjects which have never been added to the Scene before. In short,
WebScene now attempts to serialize all information on the Scene's state
with minimal simplification.